Count Autopopulation

+++ src/main/java/com/dpw/runner/shipment/services/utils/v3/ShippingInstructionUtil.java
@@ -465,7 +465,7 @@ public class ShippingInstructionUtil {
 
     private void calculateEquipmentTotals(ShippingInstructionInttraRequest response, List<CommonContainerResponse> containers) {
         int totalEquipments = containers.stream()
-                .mapToInt(container -> container.getCount() != null ? container.getCount() : 1)
+                .mapToInt(container -> Math.toIntExact(container.getCount() != null ? container.getCount() : 1))
                 .sum();
 
         int totalPackages = containers.stream()
